[4596] Move `StreamCdnResizeImageQueryParameterKeys` to the client module and make its visibility internal since it shouldn't be significant for the end user.
